Footer component moved to ts
